Added to Save Data and Results Data
Added information to the save data and the results data. The results scene hasn't been edited to add in all the new data yet.

+++ Assets/Scripts/Gameplay/GameplayManager.cs
@@ -17,7 +17,7 @@ namespace RM_MST
         public float gameTime = 0;
 
         // The game score
-        public int gameScore = 0;
+        public float gameScore = 0;
 
         // Pauses the timer if true.
         private bool gamePaused = false;
